Require query responses to be typed as a JSONValue
